Fix access token deletion: keep public repos, remove only private ones after access token deletion and confirmation message for the deletion added

+++ BTP_FE/src/app/components/get-access-tokens/get-access-tokens.component.ts
@@ -36,15 +36,19 @@ export class GetAccessTokensComponent implements OnInit {
     });
   }
 
-  deleteAccessTokenById(id: number) {
-    this.ds.deleteAccessTokeById(id).subscribe({
-      next: () => {
-        alert('Access token deleted successfully');
-        this.loadAccessTokens();
-      }, error: (err) => {
-        console.error('Error deleting access token, ', err);
-      }
-    });
+  deleteAccessTokenById(id: number, identifier: string) {
+    const confirmation = confirm(`Are you sure you want to delete ${identifier}?\nAll associated PRIVATE repositories will also be deleted.`);
+
+    if (confirmation) {
+      this.ds.deleteAccessTokeById(id).subscribe({
+        next: () => {
+          alert(`Access Token ${identifier} and associated private repositories deleted.`);
+          this.loadAccessTokens();
+        }, error: (err) => {
+          console.error('Error deleting access token, ', err);
+        }
+      });
+    }
   }
 
   getRepositoryURLs(repositories: RepositoryURLDTO[]): string {
